Move `/scat-cs` to app router
